+++ src/app/service/poap/DistributePOAP.ts
@@ -1,19 +1,18 @@
 import { AwaitMessagesOptions, DMChannel, GuildMember, Message, MessageAttachment } from 'discord.js';
 import axios from 'axios';
-import POAPUtils, { FailedPOAPAttendee, POAPFileParticipant } from '../../utils/POAPUtils';
+import POAPUtils, { POAPFileParticipant, TwitterPOAPFileParticipant } from '../../utils/POAPUtils';
 import ValidationError from '../../errors/ValidationError';
 import { Db } from 'mongodb';
 import constants from '../constants/constants';
 import { CommandContext } from 'slash-create';
 import Log, { LogUtils } from '../../utils/Log';
 import { Buffer } from 'buffer';
-import { getBufferForFailedParticipants } from './EndPOAP';
 import MongoDbUtils from '../../utils/MongoDbUtils';
 import ServiceUtils from '../../utils/ServiceUtils';
 
-export default async (ctx: CommandContext, guildMember: GuildMember, type: string, event: string): Promise<any> => {
+export default async (ctx: CommandContext, guildMember: GuildMember, event: string, platform: string): Promise<any> => {
 	if (ctx.guildID == undefined) {
-		await ctx.send('Please try ending poap event within discord channel');
+		await ctx.send('Please try poap distribution within discord channel');
 		return;
 	}
 	Log.debug('starting poap distribution...');
@@ -22,23 +21,44 @@ export default async (ctx: CommandContext, guildMember: GuildMember, type: strin
 	POAPUtils.validateEvent(event);
 	
 	await ServiceUtils.tryDMUser(guildMember, 'Hi, just need a moment to stretch before I run off sending POAPS...');
-	const participantsList: POAPFileParticipant[] | FailedPOAPAttendee[] = await askForParticipantsList(guildMember, type);
-	await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
-	let failedPOAPsList: FailedPOAPAttendee[];
-	if (type == 'MANUAL_DELIVERY') {
-		const linksMessageAttachment: MessageAttachment = await askForLinksMessageAttachment(guildMember);
-		failedPOAPsList = await POAPUtils.sendOutPOAPLinks(guildMember, participantsList as POAPFileParticipant[], linksMessageAttachment, event);
-	} else {
-		failedPOAPsList = await POAPUtils.sendOutFailedPOAPLinks(guildMember, participantsList as FailedPOAPAttendee[], event);
+	await ctx.send('Sent you a DM!');
+	let participantsList: POAPFileParticipant[] | TwitterPOAPFileParticipant[] = await askForParticipantsList(guildMember);
+	const numberOfParticipants: number = participantsList.length;
+	
+	if (numberOfParticipants <= 0) {
+		await ctx.send('Hmm there doesn\'t seem to be any participants');
+		return;
 	}
-	const failedPOAPsBuffer: Buffer = getBufferForFailedParticipants(failedPOAPsList);
+
+	if (platform == constants.PLATFORM_TYPE_TWITTER) {
+		// await distributeTwitterFlow();
+		await ctx.send('Coming soon...');
+		return;
+	}
+
+	participantsList = participantsList as POAPFileParticipant[];
+	
+	if (!participantsList[0].discordUserId) {
+		await guildMember.send({ content: 'parsing failed, please try a csv file with headers discordUserId' });
+		throw Error('failed to parse');
+	}
+
+	let failedPOAPsList: POAPFileParticipant[];
+	if (!participantsList[0].poapLink) {
+		const linksMessageAttachment: MessageAttachment = await askForLinksMessageAttachment(guildMember);
+		const listOfPOAPLinks: string[] = await POAPUtils.getListOfPoapLinks(guildMember, linksMessageAttachment);
+		failedPOAPsList = await POAPUtils.sendOutPOAPLinks(guildMember, participantsList, event, listOfPOAPLinks);
+	} else {
+		failedPOAPsList = await POAPUtils.sendOutPOAPLinks(guildMember, participantsList, event);
+	}
+	const failedPOAPsBuffer: Buffer = ServiceUtils.generateCSVStringBuffer(failedPOAPsList);
 	await guildMember.send({
 		embeds: [
 			{
 				title: 'POAPs Distribution Results',
 				fields: [
-					{ name: 'Attempted to Send', value: `${participantsList.length}`, inline: true },
-					{ name: 'Successfully Sent', value: `${participantsList.length - failedPOAPsList.length}`, inline: true },
+					{ name: 'Attempted to Send', value: `${numberOfParticipants}`, inline: true },
+					{ name: 'Successfully Sent', value: `${numberOfParticipants - failedPOAPsList.length}`, inline: true },
 					{ name: 'Failed to Send', value: `${failedPOAPsList.length}`, inline: true },
 				],
 			},
@@ -49,42 +69,24 @@ export default async (ctx: CommandContext, guildMember: GuildMember, type: strin
 		Log.debug('all poap successfully delivered');
 		return;
 	}
-	await POAPUtils.setupFailedAttendeesDelivery(guildMember, failedPOAPsList, event, ctx);
+	await POAPUtils.setupFailedAttendeesDelivery(guildMember, failedPOAPsList, event, constants.PLATFORM_TYPE_DISCORD, ctx);
 	Log.debug('poap distribution complete');
 	return;
 };
 
-export const askForParticipantsList = async (guildMember: GuildMember, type: string): Promise<POAPFileParticipant[] | FailedPOAPAttendee[]> => {
-	const message: Message = await guildMember.send({ content: 'Please upload delivery .csv file. POAPs will be distributed to these degens. Please make sure discordIds are included for each participant.' });
+export const askForParticipantsList = async (guildMember: GuildMember): Promise<POAPFileParticipant[] | TwitterPOAPFileParticipant[]> => {
+	const message: Message = await guildMember.send({ content: 'Please upload .csv file with header discordUserId and either durationInMinutes or poapLink. POAPs will be distributed to these degens.' });
 	const dmChannel: DMChannel = await message.channel.fetch() as DMChannel;
 	const replyOptions: AwaitMessagesOptions = {
 		max: 1,
 		time: 180000,
 		errors: ['time'],
 	};
-	let participantsList = [];
+	let participantsList: POAPFileParticipant[] | TwitterPOAPFileParticipant[] = [];
 	try {
 		const participantAttachment: MessageAttachment = (await dmChannel.awaitMessages(replyOptions)).first().attachments.first();
 		const fileResponse = await axios.get(participantAttachment.url);
-		participantsList = fileResponse.data.split('\n').map(participant => {
-			const values = participant.split(',');
-			if (type == 'MANUAL_DELIVERY') {
-				return {
-					id: values[0],
-					tag: values[1],
-					duration: values[2],
-				};
-			} else {
-				return {
-					discordUserId: values[0],
-					discordUserTag: values[1],
-					poapLink: values[2],
-				};
-			}
-		});
-		// remove first and last object
-		participantsList.shift();
-		participantsList.pop();
+		participantsList = ServiceUtils.parseCSVFile(fileResponse.data);
 	} catch (e) {
 		LogUtils.logError('failed to ask for participants list', e);
 		await guildMember.send({ content: 'Invalid attachment. Please try the command again.' });
@@ -102,4 +104,8 @@ export const askForLinksMessageAttachment = async (guildMember: GuildMember): Pr
 		errors: ['time'],
 	};
 	return (await dmChannel.awaitMessages(replyOptions)).first().attachments.first();
+};
+
+const distributeTwitterFlow = async () => {
+	return null;
 };