+++ src/app/service/poap/StartPOAP.ts
@@ -1,7 +1,7 @@
 import {
 	AwaitMessagesOptions,
 	Collection as DiscordCollection, DMChannel, EmbedField,
-	Guild, GuildChannel,
+	GuildChannel,
 	GuildMember, Message, MessageEmbedOptions,
 	StageChannel,
 	VoiceChannel,
@@ -26,6 +26,11 @@ import Log, { LogUtils } from '../../utils/Log';
 import dayjs, { Dayjs } from 'dayjs';
 import POAPService from './POAPService';
 import MongoDbUtils from '../../utils/MongoDbUtils';
+import VerifyTwitter from '../account/VerifyTwitter';
+import { SpaceV2LookupResult, TwitterApi } from 'twitter-api-v2';
+import apiKeys from '../constants/apiKeys';
+import { POAPTwitterSettings } from '../../types/poap/POAPTwitterSettings';
+import { POAPTwitterParticipants } from '../../types/poap/POAPTwitterParticipants';
 
 export default async (ctx: CommandContext, guildMember: GuildMember, platform: string, event: string, duration?: number): Promise<any> => {
 	if (ctx.guildID == undefined) {
@@ -42,7 +47,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 	Log.debug('poap start validated');
 	
 	if (platform == constants.PLATFORM_TYPE_TWITTER) {
-		// await startTwitterPOAPFlow(ctx, guildMember, db, event);
+		// await startTwitterPOAPFlow(ctx, guildMember, db, event, duration);
 		await ctx.send('Coming soon...');
 		return;
 	}
@@ -59,10 +64,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 		throw new ValidationError(`Please end the active event \`${activeSettings.voiceChannelName}\`.`);
 	}
 	
-	await ServiceUtils.tryDMUser(guildMember, 'Oh yea, time for a POAP event!...');
-	await guildMember.send({
-		content: 'For which voice channel should the POAP event occur?',
-	});
+	await ServiceUtils.tryDMUser(guildMember, 'Hello! For which voice channel should the POAP event occur?');
 	const voiceChannels: DiscordCollection<string, VoiceChannel | StageChannel> = ServiceUtils.getAllVoiceChannels(guildMember);
 	const message: Message = await guildMember.send({ embeds: generateVoiceChannelEmbedMessage(voiceChannels) });
 	await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
@@ -79,27 +81,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 		throw new ValidationError(`\`${channelChoice.name}\` is already active. Please reach out to <@${poapSettingsDoc.discordUserId}> to end event.`);
 	}
 	
-	const dmChannel: DMChannel = await guildMember.createDM();
-	
-	Log.debug(`asking ${guildMember.user.tag} for duration`);
-	if (duration == null) {
-		await guildMember.send({ content: 'Would you like to set the duration of the event? `(y/n)`' });
-		const setDurationFlag = (await ServiceUtils.getFirstUserReply(dmChannel)) == 'y';
-		Log.debug(`isManualEnd: ${setDurationFlag}`);
-		if (setDurationFlag) {
-			duration = await askForEventMinutes(guildMember, dmChannel);
-		} else {
-			Log.debug('max duration set for poap event');
-			duration = constants.POAP_MAX_DURATION_MINUTES;
-		}
-		Log.debug(`poap event duration: ${duration}, `, {
-			indexMeta: true,
-			meta: {
-				discordId: guildMember.guild.id,
-				discordUserId: guildMember.guild.id,
-			},
-		});
-	}
+	duration = await POAPUtils.askForDuration(guildMember, duration);
 	const currentDate: Dayjs = dayjs();
 	const currentDateISO: string = currentDate.toISOString();
 	const endDateISO: string = currentDate.add(duration, 'minute').toISOString();
@@ -128,7 +110,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 		returnDocument: 'after',
 	});
 	
-	await storePresentMembers(guildMember.guild, db, channelChoice);
+	await storePresentMembers(db, channelChoice);
 	
 	POAPService.setupAutoEndForEvent(guildMember.client, activeEvent.value);
 	
@@ -141,6 +123,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 					{ name: 'Organizer', value: `${guildMember.user.tag} `, inline: true },
 					{ name: 'Discord Server', value: `${guildMember.guild.name} `, inline: true },
 					{ name: 'Location', value: `${channelChoice.name} `, inline: true },
+					{ name: 'Platform', value: `${platform}`, inline: true },
 					{ name: 'Duration', value: `${duration} minutes`, inline: true },
 				],
 			},
@@ -186,7 +169,7 @@ export const clearPOAPParticipants = async (db: Db, guildChannel: GuildChannel):
 	});
 };
 
-export const storePresentMembers = async (guild: Guild, db: Db, channel: GuildChannel): Promise<any> => {
+export const storePresentMembers = async (db: Db, channel: GuildChannel): Promise<any> => {
 	try {
 		channel.members.forEach((member: GuildMember) => {
 			updateUserForPOAP(member, db, channel, true);
@@ -266,37 +249,106 @@ export const askUserForChannel = async (guildMember: GuildMember, dmMessage: Mes
 	}
 };
 
-const askForEventMinutes = async (guildMember: GuildMember, dmChannel: DMChannel): Promise<number> => {
-	try {
-		await guildMember.send({ content: `How long should the event stay active? \`(max: ${constants.POAP_MAX_DURATION_MINUTES} minutes)\`` });
-		const durationOfEventInMinutes: string = await ServiceUtils.getFirstUserReply(dmChannel);
-		const duration = Number(durationOfEventInMinutes);
-		POAPUtils.validateDuration(duration);
-		return duration;
-	} catch (e) {
-		LogUtils.logError('failed to process duration time', e);
-		throw new ValidationError('Please try another duration amount.');
-	}
-};
+const startTwitterPOAPFlow = async (ctx: CommandContext, guildMember: GuildMember, db: Db, event: string, duration?: number): Promise<any> => {
+	Log.debug('starting twitter poap flow...');
+	const { twitterUser } = await VerifyTwitter(guildMember);
+	const twitterClientV2: TwitterApi = new TwitterApi(apiKeys.twitterBearerToken);
+	await ServiceUtils.tryDMUser(guildMember, 'Oh yea, time for a POAP event!...');
 
-// const startTwitterPOAPFlow = async (ctx: CommandContext, guildMember: GuildMember, db: Db, event: string): Promise<any> => {
-// 	const { twitterUser, twitterClientV1 } = await VerifyTwitter(guildMember);
-// 	const twitterClientV2: TwitterApi = new TwitterApi(apiKeys.twitterBearerToken);
-// 	await ctx.send({ content: 'DM sent!' });
-// 	await guildMember.send({ content: 'Now starting poap event for twitter spaces :bird:' });
-//
-// 	// let twitterSpaceResult: SpaceV2LookupResult;
-// 	// try {
-// 	// 	twitterSpaceResult = await twitterClientV2.v2.spacesByCreators(twitterUser.id_str);
-// 	// 	console.log(twitterSpaceResult);
-// 	// 	console.log(twitterSpaceResult.includes.users);
-// 	// } catch (e) {
-// 	// 	console.log(e);
-// 	// 	LogUtils.logError('failed trying to get twitter spaces', e);
-// 	// }
-//
-// 	// if (twitterSpaceResult.meta.result_count == 0) {
-// 	// 	await guildMember.send({ content: 'Please start twitter spaces before starting POAP event.' });
-// 	// }
-//
-// };
+	let twitterSpaceResult: SpaceV2LookupResult;
+	try {
+		twitterSpaceResult = await twitterClientV2.v2.spacesByCreators(twitterUser.id_str);
+	} catch (e) {
+		LogUtils.logError('failed trying to get twitter spaces', e);
+	}
+
+	if (twitterSpaceResult.meta.result_count == 0 || twitterSpaceResult.data == null || twitterSpaceResult.data[0]['state'] != 'live') {
+		Log.warn('Twitter space result is not live');
+		await guildMember.send({ content: 'Uh-oh, please start twitter spaces before starting POAP tracking. If you have already started it, please wait a minute or two before trying again.' });
+		await ctx.send('Nothing to see here, carry on 😅... ');
+		return;
+	}
+	
+	const twitterSpaceId: string = twitterSpaceResult.data[0]['id'];
+	Log.debug(`twitter spaces event active: ${twitterSpaceId}`);
+	await ctx.send({ content: `Something really special is starting...:bird: https://twitter.com/i/spaces/${twitterSpaceId}` });
+	
+	const poapTwitterSettings: Collection<POAPTwitterSettings> = db.collection(constants.DB_COLLECTION_POAP_TWITTER_SETTINGS);
+	const activeSettings: POAPTwitterSettings = await poapTwitterSettings.findOne({
+		discordUserId: guildMember.id,
+		discordServerId: guildMember.guild.id,
+		isActive: true,
+	});
+	
+	if (activeSettings != null) {
+		Log.debug('unable to start twitter event due to active event');
+		await guildMember.send('Looks like you are already tracking your twitter space!');
+		throw new ValidationError('Ha ha.. just kidding...');
+	}
+	duration = await POAPUtils.askForDuration(guildMember, duration);
+	
+	Log.debug('setting up active twitter event in db');
+	const currentDate: Dayjs = dayjs();
+	const endTimeISO: string = currentDate.add(duration, 'minute').toISOString();
+	const twitterSettingsResult: FindAndModifyWriteOpResultObject<POAPTwitterSettings> = await poapTwitterSettings.findOneAndReplace({
+		discordUserId: guildMember.id,
+		discordServerId: guildMember.guild.id,
+	}, {
+		isActive: true,
+		event: event,
+		startTime: currentDate.toISOString(),
+		endTime: endTimeISO,
+		discordUserId: guildMember.id,
+		discordUserTag: guildMember.user.tag,
+		discordServerId: guildMember.guild.id,
+		twitterUserId: twitterUser.id_str,
+		twitterSpaceId: twitterSpaceId,
+	}, {
+		upsert: true,
+		returnDocument: 'after',
+	});
+	
+	if (twitterSettingsResult.ok != 1) {
+		Log.warn('failed to insert twitter settings active event in db for poap organizer');
+		throw new ValidationError('I\'m sorry something is not working, can you try again?');
+	}
+	
+	Log.debug(`twitter poap event stored in db and set to active for userID: ${guildMember.id}, discordServerId: ${guildMember.guild.id}`);
+	
+	POAPService.setupAutoEndForEvent(guildMember.client, twitterSettingsResult.value);
+	const claimURL = `${apiKeys.twitterClaimPage}/${twitterSpaceId}`;
+	await guildMember.send({
+		embeds: [
+			{
+				title: 'Twitter Event Started',
+				fields: [
+					{ name: 'Event', value: `${event}`, inline: true },
+					{ name: 'Organizer', value: `${guildMember.user.tag}`, inline: true },
+					{ name: 'Discord Server', value: `${guildMember.guild.name}`, inline: true },
+					{ name: 'Platform', value: 'Twitter', inline: true },
+					{ name: 'Duration', value: `${duration} minutes`, inline: true },
+					{ name: 'POAP Participation Claim Link', value: claimURL, inline: false },
+				],
+			},
+		],
+	});
+	await guildMember.send({
+		content: `POAP tracking setup! Please hand out ${claimURL} to your participants!`,
+	});
+	const poapTwitterParticipants: Collection<POAPTwitterParticipants> = db.collection(constants.DB_COLLECTION_POAP_TWITTER_PARTICIPANTS);
+	const result: FindAndModifyWriteOpResultObject<any> = await poapTwitterParticipants.findOneAndReplace({
+		twitterSpaceId: twitterSpaceId,
+		twitterUserId: twitterUser.id_str,
+	}, {
+		twitterSpaceId: twitterSpaceId,
+		twitterUserId: twitterUser.id_str,
+		dateOfTweet: currentDate.toISOString(),
+	}, {
+		upsert: true,
+	});
+	
+	if (result.ok != 1) {
+		await guildMember.send('One more thing, there was a problem setting counting you as a participants.. You just might miss the POAP, sorry.');
+	}
+	Log.debug('POAP Twitter spaces event tracking message sent');
+};