+++ src/app/types/poap/POAPTwitterSettings.ts
@@ -2,11 +2,12 @@ import { Collection, ObjectId } from 'mongodb';
 
 export interface POAPTwitterSettings extends Collection {
 	_id: ObjectId,
-	eventTitle: string,
+	event: string,
 	isActive: boolean,
 	startTime: string,
 	endTime: string,
 	discordUserId: string,
+	discordUserTag?: string,
 	discordServerId: string
 	twitterUserId: string,
 	twitterSpaceId: string,